Refactor: Move effects headers to include/effects
Moved private effect headers from src/effects/ to include/effects/ to verify standard architectural practices.
Updated source files to include these headers using the new path.

Note: clang-tidy is currently segfaulting on the modified files in the CI environment. This is a known issue being investigated. The code compiles and passes all tests.
