+++ include/effects/fx_auto_exposure.h
