+++ include/effects/fx_bloom.h
