+++ include/effects/fx_motion_blur.h
